+++ kinopoisk_unofficial/response/films/film_search_by_filters_response.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass, field
 from typing import List
 
 from kinopoisk_unofficial.contract.response import Response
-from kinopoisk_unofficial.model.dictonary.found_film import FoundFilm
+from kinopoisk_unofficial.model.found_film import FoundFilm
 
 
 @dataclass(frozen=True)
