+++ kinopoisk_unofficial/response/films/search_by_keyword_response.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass, field
 from typing import List
 
 from kinopoisk_unofficial.contract.response import Response
-from kinopoisk_unofficial.model.dictonary.found_film import FoundFilm
+from kinopoisk_unofficial.model.found_film import FoundFilm
 
 
 @dataclass(frozen=True)
